requirements.txt
PyQt6>=6.6.0
faster-whisper>=1.0.0
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
sounddevice>=0.4.6
pyaudiowpatch>=0.2.12.6; sys_platform == "win32"
groq>=0.4.0
openai>=1.0.0,<2.0.0
google-genai>=1.0.0
pypdf>=4.2.0
python-docx>=1.1.2
pynput>=1.7.6
requests>=2.31.0
mss>=9.0.1
